fix: resolve Sachin's noted issues

+++ lib/desktop_screens_new/history_screen/widgets/desktop_file_function_list.dart
@@ -146,16 +146,21 @@ class _DesktopFileFunctionListState extends State<DesktopFileFunctionList> {
               SizedBox(width: 4),
               buildOptionButton(
                 onTap: () {
-                  File(widget.filePath).deleteSync();
-                  SnackbarService().showSnackbar(
-                    context,
-                    "Successfully deleted the file",
-                    bgColor: ColorConstants.successColor,
+                  CommonUtilityFunctions().showConfirmationDialog(
+                    () {
+                      File(widget.filePath).deleteSync();
+                      SnackbarService().showSnackbar(
+                        context,
+                        "Successfully deleted the file",
+                        bgColor: ColorConstants.successColor,
+                      );
+                      Provider.of<HistoryProvider>(
+                              NavService.navKey.currentContext!,
+                              listen: false)
+                          .notify();
+                    },
+                    'Are you sure you want to delete the file(s)?',
                   );
-                  Provider.of<HistoryProvider>(
-                          NavService.navKey.currentContext!,
-                          listen: false)
-                      .notify();
                 },
                 icon: AppVectors.icDeleteFile,
               ),
